Split out deleteAllSessionsByUserId util
