+++ src/modules/forgotPassword/resolvers.ts
@@ -1,7 +1,11 @@
 import { ResolverMap } from '../../types';
 
 const resolvers: ResolverMap = {
-  Mutation: {},
+  Mutation: {
+    forgotPasswordChange: async (_, __): Promise<GQL.IForgotPasswordChangeResponse> => {
+      return { ok: true, error: null };
+    },
+  },
 };
 
 export default resolvers;
